FIX: Missing color maps in output.
Issue is caused by ITK `NrrdIO` being unable to stringify entries in `MetaDataDictionary` not defined in the nrrd standard (and `DWMRI_b-value` is not); so an empty entry would be created in the temporary nhdr file used to transfer mrml node to the CLI.

```
DWMRI_b-value:=
modality:=DWMRI
```

That missing value caused various issues in loading the temporary file when transfering back to the MRML node, so some (but not all) values in the MRML node are missing, hiding the issue.

So it seems that ad-hoc properties like this must to be stringified before passing to `MetaDataDictonary`, at least for the nrrd writer. Similarly the nrrd reader can only dynamically read strings.

I'm not sure what change caused this to suddenly break... Best I can tell, that's been the behavior since 2007, so why was it working before!? This approach seems more robust anyway, especially given `DWMRI_b-value` is a constant `1.0`.