ENH: Flush csv file buffer for intermediate results.
Each case may take a long time to run, so the file buffer may hold a case's output for a long time. This change flushes the file buffer so that intermediate results will appear in the file. Worse, if the program is aborted the buffer would never flush and never writes intermediate results.